+++ indico/modules/events/layout/forms.py
@@ -16,7 +16,7 @@ from indico.core.config import config
 from indico.modules.events.layout import theme_settings
 from indico.modules.events.layout.util import get_css_file_data, get_logo_data, get_plugin_conference_themes
 from indico.modules.users import NameFormat
-from indico.util.i18n import _
+from indico.util.i18n import _, orig_string
 from indico.web.forms.base import IndicoForm
 from indico.web.forms.fields import EditableFileField, FileField, IndicoEnumSelectField
 from indico.web.forms.validators import HiddenUnless, UsedIf
@@ -39,7 +39,31 @@ def _get_conference_theme_choices():
     return THEMES + sorted(plugin_themes, key=lambda x: x[1].lower())
 
 
-class ConferenceLayoutForm(IndicoForm):
+class LoggedLayoutForm(IndicoForm):
+    @classmethod
+    def build_field_metadata(cls, field):
+        if field.short_name == 'name_format':
+            return {'title': orig_string(field.label.text),
+                    'default': orig_string(field.none)}
+        elif field.short_name == 'theme':
+            choices = {k if k else None: orig_string(v) for k, v in field.choices}
+            return {'title': orig_string(field.label.text),
+                    'type': 'string',
+                    'convert': lambda changes: [choices[x] for x in changes]}
+        elif field.short_name == 'timetable_theme':
+            choices = {k if k else None: v for k, v in field.choices}
+            return {'title': orig_string(field.label.text),
+                    'type': 'string',
+                    'convert': lambda changes: [choices[x] for x in changes]}
+        else:
+            return orig_string(field.label.text)
+
+    @property
+    def log_fields_metadata(self):
+        return {k: self.build_field_metadata(v) for k, v in self._fields.iteritems()}
+
+
+class ConferenceLayoutForm(LoggedLayoutForm):
     is_searchable = BooleanField(_("Enable search"), widget=SwitchWidget(),
                                  description=_("Enable search within the event"))
     show_nav_bar = BooleanField(_("Show navigation bar"), widget=SwitchWidget(),
@@ -87,7 +111,7 @@ class ConferenceLayoutForm(IndicoForm):
             raise ValidationError(_('Cannot enable custom stylesheet unless there is one.'))
 
 
-class LectureMeetingLayoutForm(IndicoForm):
+class LectureMeetingLayoutForm(LoggedLayoutForm):
     name_format = IndicoEnumSelectField(_('Name format'), enum=NameFormat, none=_('Inherit from user preferences'),
                                         description=_('Format in which names are displayed'))
     timetable_theme = SelectField(_('Timetable theme'), [DataRequired()])